services/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
python-dotenv
cryptography
psycopg2-binary  # Optional for Postgres in production
